Miscellaneous PACK UI fixes
Package procedures now display a description for each parameter. (That is: we show the name of the parameter itself, now, and not just the name of whatever packdata is being fed into the parameter.)

The context menu for the procedure pane is now hooked up, though still untested.

You can no longer set a package schedule's minute if the hour is not set.

You can no longer set idles on a package template.