PACK UI fixes
Refactored the PACK UI further, to better split up the functions that synch state and UI. Fixed how we handle enable states on some controls, based on whether we are a template or use a template.

We properly synch the PACK's owning quest and interrupt type to the dialogs for editing location- and target-type package data values.

Fixed how PackageProcedureTreeModel handles QMIs, to fix a bug wherein only the first child of the root was visible (but all children of non-top-level nodes were still visible).