Critical backend and frontend fixes w.r.t conditions; PACK UI
Critical fixes to how the condition-editing UI handles package data, to fix dangling-pointer crashes and to allow the user to properly select package data as run-on entities in the first place.

Critical fixes to the condition context code, to work around mishandled smart pointers when the owning package usess a template.

Critical fix to the PACK UI: When a package uses a template, the PackageDataModel is now capable of storing and distinguishing between the local values and the inherited values.

The PACK UI now synchs with DKConditionListPanes across the editor in real-time as you edit Package Data. There are some risks with this (you could add a Package Data, edit some other form to use it, commit the other form, and then cancel the addition to the Package), but these apply to condition-editing in general, and I think the CK may have this same edge-case. Something to look into during Sustain Phase 1; I've updated my to-do notes for that accordingly.

Missing files from a previous commit: the ability to actually open the PACK UI.

Extra safeguards on the PACK UI, to prevent the user from setting a package as its own template. I suspect this could occur if the FIC listed the package as a template, but then you unflagged it as a template... maybe. Not sure.