Started work on FURN editing
Added support for loading BSFurnitureMarkerNode blocks from a NIF. Untested.

Wrote two UI model classes for FURN forms' markers. The goal is to have all marker data imported once into FurnitureMarkerModels, which will be the model for the "Markers" tableview, while using FurnitureMarkerEntryPointsProxyModel as the model for the "Entry Points" tableview and pointing the latter model at the QMarkerIndex of an entry in the former model. All untested rn.

Added functionality to `cobb::enum_flags` to: allow overwriting with a raw integer; allow casting to a raw integer; and allow overwriting a larger enum-flags mask with a smaller enum-flags mask, provided the two masks use the same enum.