FURN work
Added Natvis for cobb::enum_flags. It's not a good helper class, but as long as I'm using it, I oughta be able to easily inspect it.

FurnitureMarkersModel now properly properly extracts the FRN block from a furniture's NIF, using a new extra-data-by-name accessor on NiObjectNET.

Identified some internal-use flags on FURN forms.