PACK: Drag and drop for procedure trees
Drag-and-drop is implemented for Package procedure trees, and works for all common cases. I still need to test the edge-case of dragging orphan nodes, and dragging into/out of orphan nodes (including dragging the root into an orphan), which will require me to hand-edit an ESP in xEdit to create a PACK with orphaned procedure tree nodes.