PACK: fixes for changing procedure/branch types
Changing a procedure or branch's type now instantly updates whatever portions of the UI and models need updating.

If you change a procedure's type, we now properly update its parameters. We keep parameters unchanged when they have the same name or type (e.g. if you change from Sit to FollowTo, the "Target" parameter is unchanged); otherwise, we default parameters to any packdata that have the same name and type; otherwise, to the first found packdata with the same type; otherwise, we create new packdata.

We also track "recently used" parameters so that if you change through multiple procedure types, parameters are intact. For example, if you change from Sit to Eat to FollowTo, "Target" in the latter should retain the value it had when the procedure was Sit. (This also helps for if you accidentally change the procedure type and then immediately change it back.)