Fixes for changing what template a PACK uses
The enable states of the "Is Template" checkbox and the "Package Template" formpicker are now properly synchronized.

When changing which template a package uses, we now re-import the template's packdata values and procedure tree, and clear out any local data that doesn't map to the template (i.e. values supplied to the previous template). This process is still incomplete: we go entirely by UID without enforcing type consistency.