Bulk-renumbering forms (ESL/BEES/etc.) + non-BEES saves are safer now
Added menu bar tools to compact form IDs and to move form IDs outside of the "extended ESL" range.

If you attempt to save a v1.70 SSE file, we now catch the case of forms being in the "extended ESL" range. If possible, we'll offer to let you move them; if you refuse, or if that isn't possible, the save halts with an error.

Similarly, if you attempt to save an ESL file with forms outside of the ESL form ID range, we now catch that early and offer to let you renumber those forms in bulk, provided there aren't too many forms total to fit in the ESL form ID range. We account for whether you're saving a 1.70 or 1.71 ESL when we run these checks and operations.

When `file_load_order` searches for an available form ID, it now attempts to avoid none-stubs that it isn't capable of deleting; it should no longer treat their form IDs as available. (This change is flaky and brittle due to how the relevant checks need to be run, but fortunately, all of the code involved is fossilized enough that we shouldn't have to change things on either side. Still, this is definitely the sort of thing that I'll be refactoring the hell out of during sustain.)

Fixed how form stubs compute their cell block and sub-block numbers, to account for cells defined in an ESL. Essentially, I'm here avoiding the mistake that Bethesda made in SSE.