Fix for renumbering ESL forms and injecting out of bounds
Fixed a bug where you could renumber the active file's forms to inject them into a non-existent master, if the active file was a light file (i.e. ESL). This occurred because in that case, the destination form ID wouldn't be greater than the maximum allowed form ID (since light plug-ins go in the FE-prefixed slot).