Fixed spurious failure on conversion to ESL
DovahKit was properly converting SSE ESPs to ESL-flagged ESPs, but was then claiming that it had failed. This was because when we remapped forms' IDs from normal space to ESL space (i.e. FE-prefixed), we didn't update a map of IDs to stubs in the file-writer object, which we later used to find unsaved stubs (e.g. SSE-only forms during an SSE-to-LE save) and purge them from memory. We'd mistakenly think that all active-file-defined forms were unsaved, but we'd also fail to delete them due to what seem to be unrelated issues, causing DovahKit to warn the user about a failed save and then forcibly unload all editor content (the failsafe for when saves fail).