Major fixes to file saving; started on UI Idle tree nodes
Fixed two major bugs relating to saving files.

The first bug was that SSE-only forms weren't being stripped out of output files when converting from SSE to LE. This was due to a missing check in `file_writer`.

The second bug was that after doing an SSE-to-LE conversion, DovahKit would display spurious error messages and unload all data. This was because the tail end of the file-save process has to delete all unsaved forms (e.g. forms of SSE-only types like VOLI), and it relied on a `form_deletion_request` to do so. This was a multi-layered mistake. First: form deletion requests rely on `form_stub::load` (if a form is being deleted, we have to load its using forms into memory so we can sever those uses), which is prohibited during file-load and file-save; I resolved this by adding a passkeyed constructor to the request class for use by the file save process, and updating form stub passkeys to allow form deletion requests with the appropriate credentials to load during a file-save operation. Second: if the loading of a form triggered a warning, and that warning referenced the to-be-deleted form, then the warning object would contain a dangling pointer (because due to the nuances of Qt's signal/slot/event system, the Log Window would receive the warning object late, *after the deletion was finished), causing crashes; this was resolved by making it so that the form-load interface suppresses all warnings if the form is being loaded during a file-load or file-save operation.

This all feels very janky -- the kind of thing we'll want to freshen up during sustain -- but the problem is now fully resolved.

---

Separately from that: started work on types for the UI tree model that we'll need for editing idle animation forms.