Switching to a backend datastore for the idle tree
Documented how the game constructs the idle tree.

The way the game constructs the idle tree is complex enough that handling the logic should be done solely within the backend; so, I've started the process of rewriting the datastore as a backend object, which (when finished) should provide enough callbacks for the frontend to wrap it in a QAbstractItemModel. For now, only the initial tree-build is implemented, and still needs testing.

An initial attempt at this (dovah::utils::idle_tree_builder) was templated, but it feels more sensible to just implement the whole datastore in the backend. If the frontend needs to cache things (e.g. QString-format strings), it can store an unordered map of node pointers to cached data.