Started work on a proper Idles dialog
Moved debug code for the backend-provided datastore (and debug UI) to IdleAnimationsDialog.

The context menu option to add an action root now pops a modal defined in Qt Designer, rather than one thrown together in code. Additionally, if an action root is successfully created, it is auto-selected in the treeview.