BSA and idle-editing fixes; idle-editing work
Fixed a bug with BSA loading that caused us to break the loading of filenames.

Trying and failing to create an idle via IdleAnimationFormsModel now throws, and IdleAnimationsDialog now catches and displays an appropriate error message.

Added context menu items to IdleAnimationsDialog for adding a behavior graph and deleting an idle. Updated the context menu to show and hide menu items as appropriate based on what you right-clicked.

Game file paths as handled in the UI now de-duplicate directory separators when constructed.

The backend-provided datastore now does graph lookups by path in a case-insensitive, directory-separator-insensitive manner. This is still not a "perfect" path lookup -- redundant directory separators, for example, compare as unequal to non-redundant ones -- but it works well enough for what we're doing right now.

IdleAnimationsDialog now allows editing of a selected idle form's contents.