Condition list widgets now display enum values' backend names
When a DKConditionList is displaying conditions like IsFurnitureEntryType, where the argument types are enums, we now use the enum member names defined in the backend. These aren't localized but are an improvement over showing the raw values.

Additionally, some older changes are now synchronized into the (unused) copy of the model code in the Qt Designer plug-in source.