Idle Animations dialog is now "official"
The Idle Animations dialog is now opened via a menu bar item in the main window, and owned by DovahKitCore. Attempting to open the editor dialog for any IDLE will instead (re)open the Idle Animations dialog and focus the target IDLE within it.

DovahKitCore now offers a direct accessor to its FLO, removing the need for the "get it via the files list" hack.