Idles datastore fixes
Fixed how the idles datastore handles loose idles in a graph (we were previously loading them as loose children of the entire datastore, rather than loose idles within the graph).

When idles are drag-moved to become children of other idles, the "parent" flag on the former, if present, is cleared.