We now preserve some malformedness of IDLE on save
The degenerate data model for idle trees means that a malformed active-file IDLE record with multiple ANAM subrecords can end up placing the idle as multiple action roots, potentially in addition to making the idle loose within its canonical graph or the child of another idle. We now go out of our way to preserve this state of affairs on save, because that simplifies the handling of this case during editing (i.e. the datastore for idles doesn't have to handle the loading of multiple ANAMs, anticipate when only one ANAM would be saved, and manually make changes to the loaded hierarchy to match the effect that saving just one ANAM would have; instead the datastore can trust that if it makes no changes, the original (malformed) hierarchy is saved unchanged).