Idle datastore JS prototype: action root displacement
Added a testcase for a later-loading action root displacing an earlier-loading action root.

Fixed how `Action.prototype._track_candidate` updates the current winning root. Previously, by mistake it used the first root, not the most recently loaded root (which in this simplified JS test environment is just the last-processed root).

We now run the `#place_action_root` step when initially building idles, and actions and [action root] idles now track each other bidirectionally. The effect of this is that by the time we go to `#place_child_idle`, we already know whether the idle being placed is the winning root of any action in its canonical graph (and in fact, we can check that more accurately, particularly when the idle *is* displaced).