Idle datastore JS prototype: cross-graph displacement test
Added a testcase for displacement of an action root after it has been moved across graphs, such that it remains an action root in its former graph and becomes a loose idle in its new graph.

The datastore viewer now hides the top-level "ORPHANS" node if it is empty. (We don't hide "LOOSE" nodes because users should be allowed to create loose idles.)