Idle datastore JS prototype: detecting multiply-placed idles
When rendering the datastore out, we can now detect when the idle is multiply-placed. We accurately track an idle's canonical parent via `Idle.live.parent`, and when we're rendering an idle via some other parent, we color it in red.

Theoretically this is sufficient for a QAbstractItemModel. Thinking about it, if we refuse to render children of an idle when we've arrived at that idle via a non-canonical parent, then we may not even need "clone" nodes. Recall that a QModelIndex relies on a pointer to the parent, and a row index within the parent, to identify any given tree node.