Idle datastore JS prototype: added testcases for more degeneracies
Added testcases for the following kinds of degenerate idle trees:

An idle is placed as an action root in a master file, and relocated by the active file to be the child of another idle. (We test this occurring both wholly within one graph, and across two graphs.)

A group of idles form a cyclical previous-sibling list.

A group of idles are siblings but specify different parent idles.

An idle is placed as two action roots by a malformed active file (either one IDLE record with two DNAM+ANAM subrecords, or two IDLE records for the same idle).

An idle is placed as both an action root, and the child of another idle, by a malformed active file.

An idle is placed as one action root by one master, moved to another action root by another master, and then moved to a third action root by the active file.