Idle datastore JS prototype: "parent" flag testcase
Corrected handling of the flag in IDLE/DATA that xEdit currently calls "Parent." We no longer cache it on Idle (or rather, we no longer forget to); we just pull it directly from IdleForm, which is what the old C++ implementation does anyway.

I suspect that that flag actually means "Loose," and have added a testcase for what happens if an idle both has the flag and is an action root. I still need to test this in the CK, but with the currently implemented datastore logic in our JS prototype, the idle does indeed end up in multiple places.