Idle datastore JS prototype: verified testcases in the CK
Used xEdit to craft ESP files that match our testcases. Loading them in the CK confirms that the behaviors I expected are indeed the behaviors that occur.

Loading in the CK also indicates that a loose idle that isn't flagged as loose will trigger a warning. I've added a matching warning to the JS prototype.