Idle datastore JS prototype: "move" operation + validation diff
Implemented the algorithm for moving idles within a tree, with the appropriate callbacks fired by the datastore.

Implemented the diff viewer: the datastore element can export all forms, so we can spawn a second viewer and use it to verify that the tree we produce would be reloaded with identical content.

The datastore element can now disable drag-and-drop editing.

Fixed a mistake that caused actions and idles to track candidacies backwards (i.e. switching active-file and master-file candidacies around).

Diverged from the originally planned algorithm both in how we handle runner-up idles, and in enforcing additional preconditions (i.e. not allowing no-op moves, since that's easier than making sure we actually "move the idle to itself" properly; not allowing the movement of a master winning root into the loose idle list of that action's containing graph, since that's unrepresentable given the file format and tree-building algorithm).