Idle datastore JS prototype: idle-to-idle move fixes
Fixed the datastore-element so that you can now drop idles into other idles.

Fixed minor syntax mistakes in the code for moving one idle into another.

The datastore no longer considers it illegal to move an action root into a loose list, since the "forced loose" flag makes that possible when it would otherwise break. However, we do now explicitly prohibit moving an idle into itself or any of its own descendants, and we have another getter for moves that are technically possible but ill-advised (e.g. because they'd lead to degenerate trees).

Fixed the place-forced-loose-idle code to create graph nodes when they don't yet exist, consistent with the CK's loader.

When moving an idle across graphs, we now force updates to the DNAMs of all of its descendants.