Idle datastore JS prototype: deletion and movement fixes
Deleting an idle from its canonical-parent action now updates the latter's winning root. This is the same "update moved-from action" logic as with moving an idle, so that's now been factored out into another function.

Deleting an idle now recursively deletes its descendants. Note that this will cause an assertion failure if an idle defined in a master has been moved inside of an active-file idle that is subsequently deleted.

If an idle is implicitly loose and is explicitly moved to LOOSE, it is now made explicitly loose, rather than the move operation being considered wholly a no-op and canceled.

Deleting idles that are defined in master files is now allowed. This is necessary because you can drag-move a master-file idle to be a descendant of an active-file idle, and then delete the latter.