Idle datastore JS prototype: implementation and move-algorithm fixes
Actions now properly track the ActionForms from which they were constructed; this is needed for Idles to update their IdleForms properly.

Fixed a particular datastore operation in the "idle in active idle and active root" testcase. If you move HumanDeathRoot to become the root of ActionActivate, displacing the multiply-present HumanActivateRoot, then we now detect that both idles will end up fighting for the same action root within the active file. There's no way for us to predict or control ordering (and thus no way to reliably know which of the two will win), so we forcibly move HumanActivateRoot to loose (i.e. a full move, recursively invoking the algorithm) to avoid the conflict.

(The repro steps for that problem case were to load the "typical tree in masters" testcase, and then move HumanActivateRoot to displace DogActivateRoot.)

Fixed another problem case that could occur when displacing action roots via a recursive move. (We need to come up with terminology to distinguish "moving a node as bookkeeping" from "moving a node as an editing operation." I am here referring to the latter.) The recursive move could summon a runner-up which would then also need to be displaced.

(The repro steps for that problem case were to load the "typical tree in masters" testcase. Move HumanActivateRoot to displace DogActivateRoot, and then move HumanDeathRoot to displace HumanActivateRoot. The diff would show a correct hierarchy, but the live view would show both DogActivateRoot and HumanDeathRoot as the roots (child LI elements) of ActionActivate in Dog.hkx.)