Idle datastore: (more) consistent idle processing order
We were getting inconsistent results between the JS prototype and the C++ implementation for the "forced-loose sibling" testcase. Sometimes, but only sometimes, HumanActivate03 would appear before HumanActivate01 in its parent. This is because we were processing idles in a non-deterministic order, which works if the idle tree is entirely valid, but which causes inconsistent sibling ordering and/or error handling when there are invalid idles (such as an idle that has a correct hierarchy placement but is also flagged as loose).

The CK queues idles to load upon seeing their ANAM subrecords. As such, our approach is to sort idle nodes before we place them into the hierarchy, by checking either their action-root candidacies or (if there are none) the file pointers and offsets of the idles' form stubs. In essence, we're trying to sort idles by the order in which their first ANAM (even if in a losing record) is seen. This is almost perfect, but it will fail if an idle's losing records don't place the idle in an action root, and if the winning record is malformed and lacks ANAM; we will incorrectly treat the position/timing of the idle's winning record as the basis on which we pre-sort it. Ideally, the IDLE loader would just track the source file and offsets of the first-seen ANAM regardless of whether that indicates an action root, so we can cache that in the `idle_node` instances and query it for sorting here. So our current solution is only, like, 90% perfect.