Idle datastore debugging helpers
Added a button to the idles UI, visible only in `_DEBUG`, which does integrity checks on the datastore. All checks pass. I should probably change the inner checks to be non-member functions that are friends of the idle datastore, though, so I can move them into separate files and not have to scroll around them during normal work on the datastore code.

Added Natvis for the datastore.