DKBreadcrumbBar work; reorganized debug hooks
Reorganized several of the "debug hook" files that define items in the program's debug menu, moving them into a submenu in the UI and a subfolder in the project.

Started work on DKBreadcrumbBar, with the goal of adding one to the Idle Animations dialog. The widget is functional, but doesn't look terribly good, and certainly doesn't look close to native.