DKBreadcrumbBar work
DKBreadcrumbBar no longer use nested buttons, in favor of its own styling and paintEvent handling. This gives us much finer control over the appearance of the breadcrumb segments.

DKBreadcrumbBar now uses a minimum size hint instead of setting a fixed height when constructed.

DKBreadcrumbBar's textbox can now be opened and dismissed, though we don't properly fill its value nor check the value when dismissed; it's basically non-functional.

The arrow keys can now be used to change which segment is currently hovered. If a segment's menu is already open, then that menu is closed, and the newly-hovered segment's menu is opened. This is consistent with the Windows-native breadcrumb bar as tested on Windows 10, and is the reason we now track (by list index) which segment has a menu open. (It's also why we have to "eavesdrop" on QMenu events.)