DKBreadcrumbBar work
Implemented text editing, with the ability to configure whether it's enabled, what the separator character is, and whether paths are case-insensitive.

Implemented the ability to set a "root menu" that is shown before the segments.

Implemented "disabled" behavior and styles.

If any segments are culled and there is no root menu, we now attempt to show a "..." glyph where the root menu's button would be. However, there's a bug where if the root menu is visible, we accidentally cover it up.

Keyboard navigation within the widget can no longer hover a culled segment.

The Esc key now closes the textbox, if it's open.