DKBreadcrumbBar work
Moved per-segment repaint code to a member function on the segment struct, so that the `paintEvent` implementation can focus on the overall logic.

Redesigned how segments draw their rects, to clarify the situation with overlapping borders.

Changed the widget's focus policy to "strong focus," so that it takes focus when clicked and not merely when tabbed to.

When setting a root menu, we now listen to QMenu signals so we can properly track which of our menus is open.

Greatly simplified the layout code, and added RTL layout support.

Fixed how we hook keyboard events while menus are open: we no longer intercept them if the menu has an active action (i.e. if any of its actions have keyboard focus or have been mouseover'd).

Fixed keyboard navigation with respect to the presence of the root button.

Pressing Space or Enter while the root button has hover focus will now open the textbox, consistent with the native Windows breadcrumb bar. Pressing Space or Enter on any other segment navigates to that segment, as if you clicked it. These behaviors apply even if the segment menu is open, unless the menu has an active action (again, consistent with Windows).

Menu button icons are now cached in advance as QPainterPaths and reused, instead of rebuilt on each paint. Additionally, icon coordinates are now relative to the icon's centerpoint.

When the widget loses focus, we now forcibly close its textbox, if open, and clear its hover state. We also clear hover state when the mouse leaves the widget (unless the mouse is moving over a segment menu).