DKBreadcrumbBar fixes
DKBreadcrumbBar now behaves more faithfully to the Windows widgets when using arrow keys on the trailing segment (fixed an issue where we always opened segment 0, never the root button), when you press Esc to exit "menu mode" on the trailing segment (which as no menu), and when programmatically closing the root menu.

When a menu is opened using the vertical arrow keys, you can now press a vertical arrow key to actually move focus into the menu.