DKBreadcrumbBar: constrained-layout fixes
Fixed how we compute when segments need culling. Fixed how we truncate a segment, when that is necessary. These fixes resolve errors where a segment too large to fit in the widget would have its borders cut off, as well as fixing a crash that could occur when resizing the widget.

Fixed a crash that could occur when setting the widget's root menu to nullptr.