DKBreadcrumbBar fixes
DKBreadcrumbBar's owned menus (but notably not its root menu) are now forced to be scrollable, ensuring that they never expand to be multiple columns.

When a DKBreadcrumbBar's segment menu contains so many items as to make the menu overlap the bar, moving the mouse over the area of overlap now counts as mousing over the menu, not the bar. Previously, it counted as mousing over the bar, changing the hovered segment and effectively making the overlapped areas of the menu impossible to activate.

Improved positioning of the segment menus for RTL layout. This is still flawed: it relies on hardcoded constants for metrics that we should probably pull from QStyle (if that's possible).