DKBreadcrumbBar work and fix
Added Qt slots to control text editing. These could be useful for e.g. a custom context menu with an "Edit address" option, as is present in the breadcrumb bar in Windows 10's open/save dialog.

Added a signal for the current path changing.

Added getters for some widget states, such as how many segments are visible and whether the textbox is open.

Made some existing property getters constexpr.

Fixed a crash that could occur when keyboard-navigating to the menuless segment -- a null pointer dereference in `_open_menu` when trying to determine where on-screen to spawn a menu at.