DKBreadcrumbBar: recycles a single menu; fixes
Instead of giving every segment its own QMenu, we now just have a single QMenu that we recycle.

When we receive model updates, we now update the segment menu, if it's open. This is easier now that we just recycle a single menu. We also now account for items being moved, and our data-changed handler is actually hooked up now. These handlers are untested, mind.

Improved positioning of a segment menu when any of its items have icons.

Fixed constructor definition for DKBreadcrumbBar::Styles.