Camera Paths window
Camera Paths work very similarly to Idle Animations, except that camera path trees are much, much simpler: there's basically only one node type (the camera paths themselves), and there are no degenerate trees. The only complication is a minor one: top-level nodes are ordered, not unordered.

This means that supporting camera paths is as simple as copying the idle code and then ripping out, like, 40% of it. Of course, the need to track whole-form-type windows on DovahKitCore means this'll take ages to compile, so compilation and testing are pending.