Camera path fixes
This is needed to determine a form's position amongst its siblings if it fails to specify a previous sibling but does specify a parent. The insert-sorted-child algorithm appends the form in that scenario, before doing the after-the-fact sort, and that's the point at which the state of the list (i.e. the order in which forms are processed) matters.

Additionally, fixed a mistake in the implementation of the idles and camera-paths datastores involving forms that specify null as their previous sibling.

Fixed a datastore bug with moving camera paths, wherein we didn't remove them from their former parent list. This meant that same-parent moves would cause the same camera path to be present in the list twice.