Datastores and deleted records
If a CPTH or IDLE form has the "deleted" flag at the time the game or CK inserts it into its hierarchy parent, then it always gets appended to the end of its parent.  Presumably Bethesda did this so that "deleted"-flagged forms are (more likely to be) pushed down to the ends of their parents, effectively treating them as cruft and moving them out of the way within the CK's UI, but TESV.exe does have this behavior as well.