Started work on Story Manager support
Wrote a backend-provided datastore for the Story Manager.

Wrote an editor subsystem which wraps the backend-provided datastore, offering accessors and signals. The intention is that each "story manager event node" dialog will have its own treeview and its own model, scoped to the given event node, with that model querying the backend-provided datastore through the editor subsystem.

It may be prudent to go a little further with it, too, e.g. caching story manager node information (e.g. what event an SMEN wraps; whether an SMBN or SMQN is stacked or random; what quests an SMQN wraps; etc.) in the editor subsystem. Something to work on when I next have time.