Datastore fixes; DKScopedProxyModel
Story manager datastore fixes.

Created a general-purpose DKScopedProxyModel class, and a testcase for it. This class is a proxy model that treats an arbitrary QMI in the source model as the root node, essentially scoping its view to that root.

Out of time rn to compile; will attempt a build and run the test later