Story manager editing work
Fixes to how the SM datastore places nodes, regarding placement order.

Added a "normalize for editing" function to the SM datastore.

Created a new frontend subsystem for the story manager, and started work on the UI for editing SMENs and their descendant nodes. This UI is unfinished, but you can view the SM*N hierarchy accurately enough right now.