Story Manager work: Object Window now practical for SMENs
The Object Window now identifies SMEN forms by their event names instead of their (typically blank) editor IDs. It relies on the story manager frontend subsystem to get the SMEN event type, instead of us exposing that through the form-info-cache, so that we have a single source of truth within the frontend.

Improved the logic for how StoryManagerFormsModel caches node data. The code desperately needs cleanup, but now covers the case of a quest being edited out from under us (we seek out any SMQN nodes that may refer to it, and trigger data-changed signals in case the quest editor ID changed), and the case of an SMQN being edited but not having major changes made to its quest list (we just update the individual properties on quest entries; we don't blow away the whole list unless that's genuinely easier).