Story Manager editing work
Implemented the context menu for the SMEN dialog.

Added a breadcrumb bar to the SMEN dialog.

Refactored how we get quest properties into and out of the model. I want to avoid `setData` in order to be fully sure that I properly avoid Qt's drag-and-drop jank, *and* because it'd be inconsistent to have `setData` work for quest-form QMIs but not for any other QMI.