Work on getting SMQN UI to only allow adding legal quests
When adding quests to an SMQN, we now try to filter from the selection any quests that don't have a matching event code, and any quests already in that particular SMQN. To facilitate this, the form-info-cache subsystem now tracks quests' event codes.

(This doesn't work, because changes to the form-info-cache are incomplete. See next commit.)

Refactored how the form-info-cache subsystem handles updates to form-type-specific data, so that we can do all the relevant edits in one place. Refactored the getters to use a macro, for consistency and brevity.