Form-info-cache refactor and fixes for QUST/ENAM
Refactored the form-info-cache's internals and metaprogramming.

We now use metaprogramming and X-macros to handle form types for which we cache only one subrecord. This significantly reduces code duplication, and helps to avoid the risk of forgetting to add code to handle a given form type (that mistake being why QUST/ENAM from the previous commit didn't work).