SMEN UI now accounts for newly-created SMEN forms
Fixed a bug in the StoryManagerFormsModel that caused crashes when attempting to retrieve `data()` about newly-created SMEN forms.

Updated the SMEN UI so that when you create and edit a new SMEN form, the event type combobox is enabled but everything else is greyed out. Once you pick an event type, the form is saved and the event type is locked, consistent with the CK, and at that point you're permitted to edit the event tree.