Started work on IMAD UI support
Created UI-side type definitions for Imagespace Modifier keyframes.

Form data stores each individual imagespace property as a list of keyframes and values. The current plan for the UI is to instead store a list of keyframes each consisting of optionals for the values. In other words, the backend uses a struct-of-arrays but the frontend will use an array-of-structs. There's a lot of functionality that'll be easier this way, such as being able to have a seek slider in the UI jump from keyframe to keyframe.