Fixes for last commit; LCTN UI started
Fixed cobb::optionals::finite_float, used in the previous commit.

Adjusted the keyframe X-macro used in the previous commit, so that it requires less metaprogramming. MSVC's being flaky with if-constexpr and member access.

Implemented a UI for editing Location forms. The part of the UI dedicated to editing the form is untested. The part of the UI dedicated to showing a Location's contents isn't working properly yet.