loaded_form_ptr now casts ObjectReference subclasses to ObjectReference
This was breaking our code to gather refs for locations, but it would've probably broken plenty of other stuff as well.

The loaded_form_ptr class is part of DovahKit's oldest foundations, which is part of why it sucks, and is bad. The ergonomics involved aren't great, but for now, this change should suffice. It'll force a rebuild of the entire program, since the pointer is both widely used *and* crammed into the form stub header (I really should bite the bullet and move it to its own file eventually), so I won't be able to test this until I get home.