IMAD-related fixes
Fixed Natvis for my optional-finite-float class.

Adjusted the use of X-macros in `keyframe_collection` so that the lambda being invoked is no longer part of the macro; instead, the lambda is passed another, smaller, lambda which is used to access the X-macro'd destination field. This means we can now set breakpoints in these lambdas.

Turns out, the time values stored on interpolators in an IMAD are timeline positions, not timestamps. They're values in the range [0, 1]; to get timestamps, multiply them by the IMAD duration. The `keyframe_collection` type now accounts for this.

Fixed control flow in `keyframe_collection::get_computed_keyframe`.

With these changes, interpolation of values in an IMAD now works!