Program icon
Gave DovahKit an icon! It's a simplified version of my personal logo, inscribed in a gear that uses the same color as the Creation Kit logo.

I authored the icon as a handwritten SVG, using JavaScript to compute the gear shape. I then created a copy with the JS-generated paths substituted into the code, so I could import the result with GIMP. There, I hand-tweaked the alpha mask and some pixel-level details for the 32x32px and 16x16px sizes to improve legibility of my personal logo and (especially for the 16px size) of the gear shape.

Win32 resources are used for the icon shown when you look at the executable in Windows Explorer. The icon used for application windows is controlled via QApplication::setWindowIcon.