REFR RE/R&D, backend work, and UI work
Additional REFR R&D.

Loading a REFR now triggers some of the warnings that the CK would emit. Other warnings would require loading multiple forms in tandem, which I'm not really sure I want to do, so those are instead emitted by our REFR UI. (This isn't ideal; we'd want loading a REFR via the Render Window to potentially trigger those warnings, too; but it'll suffice for now.)

Forgotten changes from previous commits: Linked Refs and Activate Parents UI models.

REFR UI now disallows setting lock data on an automatic door, and updates multibound extents to match primitive extents.

Continuing to work to reconstruct everything that the REFR UI needs to match the CK. Water reflections are the current knot to untangle.

Added a constant in the backend to indicate the max sane bounds for an interior cell. Worldedit already respected these bounds because the game bugs out past them, but it did so with the bounds written inline. It turns out that the CK's REFR loader warns on those bounds, so now it and Worldedit use a shared constant.